SageMath version 8.3, Release Date: 2018-08-03

* tag '8.3': (8421 commits)
  Updated SageMath version to 8.3
  Updated SageMath version to 8.3.rc3
  although not strictly necessary, it would be good to bump the patch level if nothing else so the buildbots can see that the package changed
  Make absolutely sure that whatever else happened before installing or upgrading this package, we remove the old $SAGE_LOCAL/bin/pkg-config before trying to generate the new one, or lese it can become self-referential.  Fixes https://trac.sagemath.org/ticket/25912
  Updated SageMath version to 8.3.rc2
  Add a stopgap to DuadicCodeEvenPair/DuadicCodeOddPair
  Add singular as dependency for p_group_cohomology
  bump version of pkgconf
  Updated SageMath version to 8.3.rc1
  uninstall should happen before build
  Upgrade cysignals to 1.7.2
  adjust for minor system-dependent floating point variation on this test; #25815
  allow sdh_install to simply overwrite existing files when the -T option is used
  #25620 work around weakness of _arb_vec_sort_pretty in test
  Updated SageMath version to 8.3.rc0
  Fix doctest due to more verbose python message
  25529: corrected doc-build failure
  fix uninstall script
  Upgrade to Python 3.6.6
  trac 25579 fixing bad handling of error msg in plot.py
  ...

+++ src/sage/misc/inline_fortran.py
@@ -1,20 +1,21 @@
 """
 Fortran compiler
 """
+from __future__ import absolute_import
 from six import iteritems
+
 import os
 import imp
 import shutil
+import sys
 
 from sage.misc.temporary_file import tmp_dir
 
 
 class InlineFortran:
     def __init__(self, globals=None):
-        if globals is None:
-            self.globs = {}
-        else:
-            self.globs = globals  # Deprecated
+        # globals=None means: use user globals from REPL
+        self.globs = globals
         self.library_paths=[]
         self.libraries=[]
         self.verbose = False
@@ -80,6 +81,9 @@ class InlineFortran:
         """
         if globals is None:
             globals = self.globs
+            if globals is None:
+                from sage.repl.user_globals import get_globals
+                globals = get_globals()
 
         from numpy import f2py
 
@@ -128,11 +132,16 @@ class InlineFortran:
                 print(log_string)
         finally:
             os.chdir(old_cwd)
-            try:
-                shutil.rmtree(mytmpdir)
-            except OSError:
-                # This can fail for example over NFS
-                pass
+
+            if sys.platform != 'cygwin':
+                # Do not delete temporary DLLs on Cygwin; this will cause
+                # future forks of this process to fail.  Instead temporary DLLs
+                # will be cleaned up upon process exit
+                try:
+                    shutil.rmtree(mytmpdir)
+                except OSError:
+                    # This can fail for example over NFS
+                    pass
 
         for k, x in iteritems(m.__dict__):
             if k[0] != '_':
